[open source insight] Set the github action dependencies by sha
I think the version number is clearer but this is done in order to
have a better score in https://deps.dev/pypi/pylint. The reasoning
is that it mitigate the risk of compromised actions.

It seems it's not going to change anything for the dependency update by dependabot:
https://github.com/systemd/systemd/pull/22638

Full disclosure, I'm being paid to make pylint more secure by Tidelift
and this is part of this effort.