Adding optional checker for nested min/max
This adds a new checker (not active by default) which identifies usages
similar to ``min(<arg1>, min(<arg2>, <arg3>))`` and suggests using a
simplified form of ``min(<arg1>, <arg2>, <arg3>)``. Same goes for
``max`` usage.

The logic is as follows: it detects calls to either ``min`` or ``max``
functions, and whenever one of their arguments is that same function, it
emits the message.