Fix check unused arguments false positive bug
Problem: the special method `__new__` must match the arguments of the
`__init__` method even if `__new__` method does not use them. This
generate `unused-argument` for the `__new__` method.

Fix: the unused arguments check should not be done on the `__new__`
method if the `__init__` method is defined in the same class.

Fixes https://github.com/pylint-dev/pylint/issues/3670