Update `unused-argument` functional test
Update `unused-argument` test to include a check for the case of
`__init__` and `__new__` being defined in a class but `__new__` does not
use all of the argument. This is fine because `__new__` must have the
same argument of `__init__`.

Update with a second check in case of `__init__` being not defined in a
class. Then the unused arguments check must be done on `__new__`.