Fix check unused arguments false positive bug (#8542)
Problem: the special method `__new__` must match the arguments of the
`__init__` method even if `__new__` method does not use them. This
generate `unused-argument` for the `__new__` method.

Fix: the unused arguments check should not be done on the `__new__`
method if the `__init__` method is defined in the same class.

Update `unused-argument` test to include a check for the case of
`__init__` and `__new__` being defined in a class but `__new__` does not
use all of the argument. This is fine because `__new__` must have the
same argument of `__init__`.

Update with a second check in case of `__init__` being not defined in a
class. Then the unused arguments check must be done on `__new__`.

Fixes https://github.com/pylint-dev/pylint/issues/3670

(cherry picked from commit 156da64d0fb4c06e15c5b619b91ce550d594a770)