Extend test cases for 'condition-evals-to-constant'
Related to #7933.

The original issue reported pylint as missing bugs in comparisons
like:

`if get_text() == "string A" or "string B":`

In reality this is already covered by the message
'condition-evals-to-constant', however test cases with strings
were missing, they have been added to self-document other cases
this checker applies to and extend coverage.