[base checker] Fix the ordering of checkers for extensions (#8956)
The comparison was not done properly for extension when the second values
given was a builtin and the first one an extension. The tests were not
testing this case properly because the alphabetical order made us think
that everything went right. Added a test with a low alphabetical value extension checker.

Necessary for #8951