New message `use-yield-from` (R1737) (#9419)
Suggest refactoring when yielding from a for loop can be replaced by 
'yield from'. This message is only emitted when there are no other 
statements in the containing for loop.

There are too many possible generators to use in a for loop and handling
all the cases to provide a human friendly display name in emitted
messages seems too cumbersome compared to the benefit it brings,
specially considering this message will have a doc page with examples.

It is invalid to use 'yield from' in async functions. Replacing yielding
from an 'async for' loop may not be the same as using 'yield from' so
the message won't be emitted for async loops as well.

Closes #9229