📝 Integrate `sphinxcontrib-towncrier`
This plugin embeds displaying a changelog draft right into the
Sphinx ecosystem, removing the need to separately run the
`towncrier --draft` command externally.
It works with native Sphinx build runs universally, meaning that
the result is the same and embedded for local runs, and is not
only set up on RTD in a disconnected manner.

*DISCLAIMER:* I authored this Sphinx extension.