Fix a false positive for ``redefined-outer-name`` (#9678)
When there is a name defined in an exception-handling block which shares the same name
as a local variable that has been defined in a function body. Check if the outer node is in the
scope of an exception assignment and do not emit ``redefined-outer-name`` if that is the case.

Closes #9671

(cherry picked from commit 57ae027efcd9bc19d793459fc15904cb9c45578a)