[logging-too-few-args] Raise for format strings with no interpolation at all  (#10321)
* [feat] Detect missing formatting args when the string is not interpolated too

The original decision was taken in db8b3a47be during implementation with
a different rational ("If no args were supplied, then all format strings
are valid don't check any further.") and was not discussed again when the
comment was updated in #2713. I think the new behavior make sense especially
considering that the primer shows 4 false negative in home-assistant.

Co-authored-by: Alex Prabhat Bara <alexpbara@gmail.com>

* Fix required because logging are tested elsewhere