Fix caching issue that led to incorrect all-options.rst being merged on main (#10320)
Using tox works because tox does not use the cached venv from the github action.
It's not efficient but this is a fast fix. We removed the pre-commit hook for doc 
generation because it's too slow to be usable and we have already two ways to
generate the doc. Also "make build-html" does not exists anymore so upgraded the doc.

Co-authored-by: Pierre Sassoulas <pierre.sassoulas@gmail.com>