Fix two multiprocessing issues (#10642)
1) Not possible to enable extension checks with `"default_enabled": False`
Whenever `register_checker` is called it overwrites the message filter in case a message is disabled by default. Usually that isn't an issue as the checkers are first registered and the config loaded afterwards.
https://github.com/pylint-dev/pylint/blob/6ce60320a4a369323650f2d4babd57955752e411/pylint/lint/pylinter.py#L495-L504

For multiple jobs the dynamic checkers are re-registered though which overwrites the previously loaded config. To resolve that, it's necessary to restore the message state by caching the list of enabled and disabled messages and reapplying it.
https://github.com/pylint-dev/pylint/blob/6ce60320a4a369323650f2d4babd57955752e411/pylint/lint/parallel.py#L55-L58
Closes #10037

2) Warnings are emitted multiple times
Whenever a checker get's registered, it's just appended to the list of checkers with the corresponding checker name. That's necessary to allow multiple modules to register the same checker name.
https://github.com/pylint-dev/pylint/blob/6ce60320a4a369323650f2d4babd57955752e411/pylint/lint/pylinter.py#L495-L497

When the checker is now re-registered during the worker init, the "old" one is still active. Thus any message is actually printed twice. To resolve it, deregister the "old" dynamic checkers before loading them again.